#8 Skapade Project2

+++ SuperDuperClassLibrary/MathServices.cs
@@ -14,7 +14,7 @@
         {
             return num1 * num2;
         }
-        public double Division(int num1, int num2)
+        public double Division(double num1, double num2)
         {
             return num1 / num2;
         }
